+++ library/src/main/java/com/deltadna/android/sdk/DDNAImpl.java
@@ -379,6 +379,11 @@ final class DDNAImpl extends DDNA {
     public DDNA forgetMe() {
         return stopSdk();
     }
+
+    @Override
+    public DDNA stopTrackingMe() {
+        return stopSdk();
+    }
     
     @Override
     ImageMessageStore getImageMessageStore() {